Move tool call resolution into resolver module
