+++ tests/test_base.py
@@ -1,5 +1,4 @@
 from unittest.mock import AsyncMock
-from types import SimpleNamespace
 import pytest
 import json
 from omnicoreagent.core.agents.base import BaseReactAgent, TOOL_CALL_TIMEOUT_MESSAGE
@@ -685,143 +684,3 @@ async def test_handle_tool_loop_state_marks_session_stuck(agent):
     assert events[0]["event"].type == EventType.TOOL_CALL_ERROR
     assert events[0]["event"].payload.tool_name == "alpha"
 
-
-@pytest.mark.asyncio
-async def test_resolve_single_tool_action_uses_local_registry(agent):
-    registry = ToolRegistry()
-
-    @registry.register_tool(
-        name="local_ping",
-        inputSchema={
-            "type": "object",
-            "properties": {"value": {"type": "string"}},
-            "required": ["value"],
-        },
-        description="Local ping tool.",
-    )
-    async def local_ping(value: str):
-        return {"status": "success", "data": value}
-
-    resolved = await agent._resolve_single_tool_action(
-        action={"tool": "local_ping", "parameters": {"value": "runtime"}},
-        sessions={},
-        mcp_tools=None,
-        local_tools=registry,
-    )
-
-    assert isinstance(resolved, ToolCallResult)
-    assert resolved.tool_name == "local_ping"
-    assert resolved.tool_args == {"value": "runtime"}
-
-
-@pytest.mark.asyncio
-async def test_resolve_single_tool_action_uses_mcp_tools_before_local(agent):
-    registry = ToolRegistry()
-
-    @registry.register_tool(
-        name="search",
-        inputSchema={"type": "object", "properties": {}, "required": []},
-        description="Local search fallback.",
-    )
-    async def search():
-        return {"status": "success", "data": "local"}
-
-    resolved = await agent._resolve_single_tool_action(
-        action={"tool": "search", "parameters": {"query": "runtime"}},
-        sessions={"server": {"session": object()}},
-        mcp_tools={"server": [SimpleNamespace(name="search")]},
-        local_tools=registry,
-    )
-
-    assert isinstance(resolved, ToolCallResult)
-    assert resolved.tool_name == "search"
-    assert resolved.tool_args == {"query": "runtime"}
-    assert resolved.tool_executor.tool_handler.server_name == "server"
-
-
-@pytest.mark.asyncio
-async def test_resolve_single_tool_action_rejects_tool_call_to_sub_agent(agent):
-    sub_agent = SimpleNamespace(name="research_agent")
-
-    resolved = await agent._resolve_single_tool_action(
-        action={"tool": "research_agent", "parameters": {"task": "inspect"}},
-        sessions={},
-        mcp_tools=None,
-        local_tools=None,
-        sub_agents=[sub_agent],
-    )
-
-    assert isinstance(resolved, ToolError)
-    assert resolved.tool_name == "N/A"
-    assert "is a sub-agent, not a tool" in resolved.observation
-
-
-def test_parse_tool_actions_handles_empty_and_single_action(agent):
-    empty = agent._parse_tool_actions(ParsedResponse(action=True, data=None))
-    assert isinstance(empty, ToolError)
-    assert empty.observation == "Invalid tool call request: No data provided"
-
-    parsed = agent._parse_tool_actions(
-        ParsedResponse(
-            action=True,
-            data=json.dumps({"tool": "local_ping", "parameters": {"value": "one"}}),
-        )
-    )
-    assert parsed == [{"tool": "local_ping", "parameters": {"value": "one"}}]
-
-
-def test_mcp_tools_for_action_disables_mcp_only_for_tools_retriever(agent):
-    mcp_tools = {"server": [SimpleNamespace(name="search")]}
-
-    assert (
-        agent._mcp_tools_for_action(
-            {"tool": "tools_retriever", "parameters": {"query": "mail"}},
-            mcp_tools=mcp_tools,
-        )
-        is None
-    )
-    assert (
-        agent._mcp_tools_for_action(
-            {"tool": "search", "parameters": {"query": "runtime"}},
-            mcp_tools=mcp_tools,
-        )
-        is mcp_tools
-    )
-
-
-@pytest.mark.asyncio
-async def test_resolve_tool_call_request_keeps_mcp_available_after_tools_retriever(agent):
-    registry = ToolRegistry()
-
-    @registry.register_tool(
-        name="tools_retriever",
-        inputSchema={
-            "type": "object",
-            "properties": {"query": {"type": "string"}},
-            "required": ["query"],
-        },
-        description="Retrieve tools.",
-    )
-    async def tools_retriever(query: str):
-        return {"status": "success", "data": query}
-
-    resolved = await agent.resolve_tool_call_request(
-        parsed_response=ParsedResponse(
-            action=True,
-            tool_calls=True,
-            data=json.dumps(
-                [
-                    {"tool": "tools_retriever", "parameters": {"query": "search"}},
-                    {"tool": "search", "parameters": {"query": "runtime"}},
-                ]
-            ),
-        ),
-        sessions={"server": {"session": object()}},
-        mcp_tools={"server": [SimpleNamespace(name="search")]},
-        local_tools=registry,
-    )
-
-    assert not isinstance(resolved, ToolError)
-    assert [tool.tool_name for tool in resolved] == ["tools_retriever", "search"]
-    assert resolved[0].tool_executor.tool_handler.local_tools is registry
-    assert resolved[1].tool_executor.tool_handler.server_name == "server"
